temporary windows fix

+++ src/artpop/image/imager.py
@@ -582,8 +582,7 @@ class ArtImager(Imager):
             src_counts[src_counts < 0] = 0
             sky_counts = 0
             mag_error = None
-        raw_counts = self.rng.poisson(
-            (src_counts + sky_counts).astype(np.float64))
+        raw_counts = self.rng.poisson(src_counts + sky_counts)
         if self.read_noise > 0.0:
             rn = self.rng.normal(scale=self.read_noise, size=src_counts.shape)
             raw_counts = raw_counts + rn
